requirements.txt
transformers
sentence-transformers
datasets
nltk
numpy
pandas
tqdm
torchtext
pytorch-model-summary
absl-py
tensorboardX
tables